Load output data into memory
